requirements.txt
Django==1.10.2
dj-database-url==0.4.1
django-cors-middleware==1.3.1
djangorestframework==3.4.5
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
model-mommy==1.2.6
psycopg2==2.6.2
whitenoise==3.2.2
